requirements.txt
absl-py==0.13.0
aiofiles==0.5.0
alembic==1.4.3
antlr4-python3-runtime==4.8
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
asgiref==3.3.1
astor==0.8.1
astroid==2.4.2
asttokens==2.0.3
astunparse==1.6.3
attrs==21.2.0
audioread==2.1.9
autopep8==1.5.7
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
black==21.12b0
bleach==4.1.0
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
chrome-gnome-shell==0.0.0
clang==5.0
click==7.1.2
cloudpickle==2.0.0
cmake==3.21.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
curlify==2.2.1
cycler==0.10.0
Cython==0.29.26
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
dlib==19.22.1
dnspython==2.0.0
docutils==0.16
duplicity==0.8.12.0
email-validator==1.1.1
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition-models==0.3.0
fasteners==0.14.1
filelock==3.0.12
fire==0.4.0
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.12
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.4.0
google-api-core==2.2.2
google-api-python-client==1.7.11
google-auth==1.35.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.6
google-cloud-speech==2.11.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
googletrans==4.0.0rc1
grpcio==1.42.0
grpcio-status==1.42.0
gTTS==2.2.3
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2021.11.1
html5lib==1.0.1
httpcore==0.9.1
httpie==1.0.3
httplib2==0.14.0
httptools==0.1.2
httpx==0.13.3
hydra-core==1.1.1
hyperframe==5.2.0
idna==2.10
imageio==2.9.0
imantics==0.1.12
imgaug==0.4.0
importlib-resources==5.4.0
imutils==0.5.4
iopath==0.1.9
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.5.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
jsonschema==4.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
labelme2coco==0.2.0
langdetect==1.0.8
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
libcst==0.3.23
librosa==0.8.1
llvmlite==0.37.0
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mediapipe==0.8.8.1
mistune==0.8.4
monotonic==1.5
msgpack==1.0.2
multidict==4.7.6
mypy==0.761
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.5
notebook==6.4.4
numba==0.54.1
numexpr==2.7.3
numpy==1.19.5
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
omegaconf==2.1.1
opencv-contrib-python==4.5.4.58
opencv-python==4.5.2.54
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.0
pandas==1.2.3
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.1
pathspec==0.9.0
pbr==5.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.0.0
pixellib==0.7.1
platformdirs==2.4.1
playsound==1.3.0
pooch==1.5.2
portalocker==2.3.2
prometheus-client==0.11.0
prompt-toolkit==3.0.17
proto-plus==1.19.8
protobuf==3.19.1
psutil==5.5.1
ptyprocess==0.7.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.2
PyDrive==1.3.1
Pygments==2.8.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pypi-xmlrpc==2020.12.3
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyRFC3339==1.1
pyrsistent==0.18.0
pyserial==3.4
pytesseract==0.3.7
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-editor==1.0.4
python-multipart==0.0.5
pyttsx3==2.90
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.1
pyzmq==22.0.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
resampy==0.2.2
rfc3986==1.5.0
roman==2.0.0
rsa==4.0
sahi==0.8.19
schematics==2.1.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
Shapely==1.8.0
simplejson==3.16.0
six==1.15.0
sklearn==0.0
slim==0.6.2
sniffio==1.2.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.19
sqlparse==0.4.1
stevedore==3.3.0
systemd-python==234
tables==3.6.1
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
terminaltables==3.1.10
testpath==0.5.0
thonny==3.2.7
threadpoolctl==2.1.0
tifffile==2021.3.5
timedelta==2020.12.3
tokenizer==3.2.0
toml==0.10.1
tomli==1.2.3
tornado==6.1
tqdm==4.58.0
traitlets==5.0.5
typed-ast==1.4.1
typeguard==2.12.1
typing-inspect==0.7.1
typing_extensions==4.0.1
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.11.8
uvloop==0.16.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
vosk==0.3.32
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
WooCommerce==3.0.0
wrapt==1.12.1
WTForms==2.3.3
xkit==0.0.0
xlrd==2.0.1
xmljson==0.2.1
yacs==0.1.8
yarl==1.4.2
zipp==3.7.0
